# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WhyNoIPv6 API
  description: >
    The public read API of whynoipv6.com (07-api.md). Versionless root,
    snake_case wire, `{items,page,meta}` / `{points,meta}` envelopes, keyset
    cursor pagination, RFC 9457 problem+json errors. The only write path is
    the anonymous async live check (`POST /check`). Data is licensed
    CC-BY-NC-4.0.
  version: "1.0.0"
  contact:
    url: https://whynoipv6.com
  license:
    name: CC-BY-NC-4.0
    url: https://creativecommons.org/licenses/by-nc/4.0/
servers:
  - url: https://api.whynoipv6.com
# The filter composition every tier/scoped list shares with /domains
# (07 §4.4). YAML anchor — expanded per path at parse time.
...
paths:
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: The general filterable domain leaderboard
      parameters:
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/country"
        - $ref: "#/components/parameters/asn"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimBase"
        - $ref: "#/components/parameters/dimWWW"
        - $ref: "#/components/parameters/dimNS"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/dimConn"
        - $ref: "#/components/parameters/dimResources"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /domains/{host}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Domain detail with masked informational dimensions
      parameters:
        - $ref: "#/components/parameters/host"
        - name: include
          in: query
          description: "`evidence` attaches the latest scan_detail payload"
          schema: { type: string, enum: [evidence] }
      responses:
        "200":
          description: The §4.3 detail representation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DomainDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/{host}/subdomains:
    get:
      operationId: listSubdomains
      tags: [domains]
      summary: An apex's non-disabled children, host-ordered
      parameters:
        - $ref: "#/components/parameters/host"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/{host}/resources:
    get:
      operationId: listDomainResources
      tags: [resources]
      summary: "Forward resource dependencies: what this domain depends on"
      parameters:
        - $ref: "#/components/parameters/host"
      responses:
        "200":
          description: Bounded small; exact `meta.count`, trivial page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceLinkList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/{host}/changelog:
    get:
      operationId: listDomainChangelog
      tags: [changelog]
      summary: Per-domain confirmed-transition feed
      parameters:
        - $ref: "#/components/parameters/host"
        - $ref: "#/components/parameters/changelogField"
        - $ref: "#/components/parameters/changelogFrom"
        - $ref: "#/components/parameters/changelogTo"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/ChangelogList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /domains/{host}/changelog.atom:
    get:
      operationId: getDomainChangelogAtom
      tags: [changelog]
      summary: Per-domain Atom feed (latest 50)
      parameters:
        - $ref: "#/components/parameters/host"
      responses:
        "200": { $ref: "#/components/responses/AtomFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/{host}/changelog.feed.json:
    get:
      operationId: getDomainChangelogJSONFeed
      tags: [changelog]
      summary: Per-domain JSON Feed 1.1 (latest 50)
      parameters:
        - $ref: "#/components/parameters/host"
      responses:
        "200": { $ref: "#/components/responses/JSONFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/{host}/history:
    get:
      operationId: getDomainHistory
      tags: [stats]
      summary: Per-domain confirmed trajectory, changelog-reconstructed
      parameters:
        - $ref: "#/components/parameters/host"
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: >
            Reconstructed confirmed state per day; `error`/`inconsistent`
            never appear. Empty changelog → `points: []` (day-1 rule).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistoryResponse" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
  /heroes:
    get:
      operationId: listHeroes
      tags: [tiers]
      summary: "Tier preset: class=hero"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /sinners:
    get:
      operationId: listSinners
      tags: [tiers]
      summary: "Tier preset: class=sinner"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /saints:
    get:
      operationId: listSaints
      tags: [tiers]
      summary: "Tier preset: saint=true"
      parameters: *tierParams
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
...
  /countries:
    get:
      operationId: listCountries
      tags: [countries]
      summary: Country adoption leaderboard
      parameters:
        - name: sort
          in: query
          schema: { type: string, enum: [percent, v6_sites, sites], default: percent }
        - $ref: "#/components/parameters/format"
      responses:
        "200":
          description: The bounded ISO set, served whole with exact `meta.count`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CountryList" }
            text/csv:
              schema: { type: string }
        "400": { $ref: "#/components/responses/InvalidParameter" }
  /countries/{code}:
    get:
      operationId: getCountry
      tags: [countries]
      summary: Country detail
      parameters: [{ $ref: "#/components/parameters/countryCode" }]
      responses:
        "200":
          description: The §4.5 representation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Country" }
        "404": { $ref: "#/components/responses/NotFound" }
  /countries/{code}/domains:
    get:
      operationId: listCountryDomains
      tags: [countries]
      summary: The country-scoped leaderboard (≡ /domains?country=)
      parameters:
        - $ref: "#/components/parameters/countryCode"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /countries/{code}/stats:
    get:
      operationId: getCountryStats
      tags: [stats]
      summary: Country adoption time series
      parameters:
        - $ref: "#/components/parameters/countryCode"
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CountryStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
  /countries/{code}/changelog:
    get:
      operationId: listCountryChangelog
      tags: [changelog]
      summary: Per-country feed (latest-50 recent window)
      parameters: [{ $ref: "#/components/parameters/countryCode" }]
      responses:
        "200": { $ref: "#/components/responses/ChangelogList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /countries/{code}/changelog.atom:
    get:
      operationId: getCountryChangelogAtom
      tags: [changelog]
      summary: Per-country Atom feed (latest 50)
      parameters: [{ $ref: "#/components/parameters/countryCode" }]
      responses:
        "200": { $ref: "#/components/responses/AtomFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /countries/{code}/changelog.feed.json:
    get:
      operationId: getCountryChangelogJSONFeed
      tags: [changelog]
      summary: Per-country JSON Feed 1.1 (latest 50)
      parameters: [{ $ref: "#/components/parameters/countryCode" }]
      responses:
        "200": { $ref: "#/components/responses/JSONFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /asns:
    get:
      operationId: listASNs
      tags: [networks]
      summary: The hosting-ASN league table
      parameters:
        - name: sort
          in: query
          schema: { type: string, enum: [count_v6, count_total], default: count_v6 }
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Keyset-paginated on (count, number) descending.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ASNList" }
            text/csv:
              schema: { type: string }
        "400": { $ref: "#/components/responses/InvalidParameter" }
  /asns/{number}:
    get:
      operationId: getASN
      tags: [networks]
      summary: Network detail (count_v4 synthesized)
      parameters: [{ $ref: "#/components/parameters/asnNumber" }]
      responses:
        "200":
          description: The §4.6 representation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ASN" }
        "404": { $ref: "#/components/responses/NotFound" }
  /asns/{number}/domains:
    get:
      operationId: listASNDomains
      tags: [networks]
      summary: The network-scoped leaderboard (≡ /domains?asn=)
      parameters:
        - $ref: "#/components/parameters/asnNumber"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /asns/{number}/stats:
    get:
      operationId: getASNStats
      tags: [stats]
      summary: Network adoption time series (canonical count_v6/count_total wire names)
      parameters:
        - name: number
          in: path
          required: true
          description: AS number; an optional leading `AS` prefix is accepted.
          schema: { type: string }
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ASNStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: The DNS-provider league table (exact stored counters)
      parameters:
        - $ref: "#/components/parameters/format"
      responses:
        "200":
          description: The bounded curated registry, exact `meta.count`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderList" }
            text/csv:
              schema: { type: string }
        "400": { $ref: "#/components/responses/InvalidParameter" }
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: DNS-provider detail
      parameters: [{ $ref: "#/components/parameters/providerID" }]
      responses:
        "200":
          description: The §4.6 provider representation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
        "404": { $ref: "#/components/responses/NotFound" }
  /providers/{id}/domains:
    get:
      operationId: listProviderDomains
      tags: [providers]
      summary: The provider-scoped leaderboard (the path form is its own indexed scope)
      parameters:
        - $ref: "#/components/parameters/providerID"
        - $ref: "#/components/parameters/class"
        - $ref: "#/components/parameters/saint"
        - $ref: "#/components/parameters/tld"
        - $ref: "#/components/parameters/provider"
        - $ref: "#/components/parameters/hosting"
        - $ref: "#/components/parameters/flag"
        - $ref: "#/components/parameters/dimMX"
        - $ref: "#/components/parameters/rankMin"
        - $ref: "#/components/parameters/rankMax"
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/sort"
        - $ref: "#/components/parameters/fields"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/afterRank"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ScopeRequired" }
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: Campaign index
      parameters:
        - name: tag
          in: query
          description: Filter to campaigns carrying this tag; unknown tags return an empty list.
          schema: { type: string }
      responses:
        "200":
          description: Bounded curated set, exact `meta.count`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignList" }
...
  /campaigns/{uuid}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: "Composite campaign detail: metadata + paged members + adoption"
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: The §4.7 composite.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/domains:
    get:
      operationId: listCampaignDomains
      tags: [campaigns]
      summary: Campaign members, host-ordered (rank-NULL rows visible)
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/DomainList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/stats:
    get:
      operationId: getCampaignStats
      tags: [stats]
      summary: Campaign adoption time series (incl. v6_ready)
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: Confirmed-state snapshots; disabled campaigns are 404.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/changelog:
    get:
      operationId: listCampaignChangelog
      tags: [changelog]
      summary: Campaign-wide feed (latest-50 recent window)
      parameters: [{ $ref: "#/components/parameters/campaignUUID" }]
      responses:
        "200": { $ref: "#/components/responses/ChangelogList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/domains/{host}/changelog:
    get:
      operationId: listCampaignDomainChangelog
      tags: [changelog]
      summary: One member's feed (404 for non-members)
      parameters:
        - $ref: "#/components/parameters/campaignUUID"
        - $ref: "#/components/parameters/host"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/ChangelogList" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/changelog.atom:
    get:
      operationId: getCampaignChangelogAtom
      tags: [changelog]
      summary: Per-campaign Atom feed (latest 50)
      parameters: [{ $ref: "#/components/parameters/campaignUUID" }]
      responses:
        "200": { $ref: "#/components/responses/AtomFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{uuid}/changelog.feed.json:
    get:
      operationId: getCampaignChangelogJSONFeed
      tags: [changelog]
      summary: Per-campaign JSON Feed 1.1 (latest 50)
      parameters: [{ $ref: "#/components/parameters/campaignUUID" }]
      responses:
        "200": { $ref: "#/components/responses/JSONFeed" }
        "404": { $ref: "#/components/responses/NotFound" }
  /resources/{host}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Resource-host headline (dependent_count)
      parameters: [{ $ref: "#/components/parameters/host" }]
      responses:
        "200":
          description: The §4.11 headline block.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResourceHostDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
  /resources/{host}/dependents:
    get:
      operationId: listResourceDependents
      tags: [resources]
      summary: "Reverse dependents: the advocacy surface"
      parameters:
        - $ref: "#/components/parameters/host"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: >
            Null-flag-first ordering (rank NULLS LAST); items are §4.2 rows
            plus the link attributes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DependentsResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /changelog:
    get:
      operationId: listChangelog
      tags: [changelog]
      summary: Global recent-transitions feed
      description: >-
        With `scope=campaign` the feed is restricted to transitions of domains
        that are members of any campaign, served as the fixed latest-50 recent
        window (no pagination — `page` cursors are null); the other query
        parameters are ignored in that mode.
      parameters:
        - $ref: "#/components/parameters/changelogScope"
        - $ref: "#/components/parameters/changelogField"
        - $ref: "#/components/parameters/changelogFrom"
        - $ref: "#/components/parameters/changelogTo"
        - $ref: "#/components/parameters/format"
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200": { $ref: "#/components/responses/ChangelogList" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [stats]
      summary: The headline adoption dashboard series
      parameters:
        - $ref: "#/components/parameters/statsFrom"
        - $ref: "#/components/parameters/statsTo"
        - $ref: "#/components/parameters/interval"
      responses:
        "200":
          description: "`meta.source` is always `confirmed_state`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlobalStatsSeries" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
  /datasets:
    get:
      operationId: getDatasets
      tags: [datasets]
      summary: The static-dataset manifest
      responses:
        "200":
          description: Re-read from disk per request; files are nginx-served.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Manifest" }
        "503": { $ref: "#/components/responses/ManifestUnavailable" }
  /check:
    post:
      operationId: createCheck
      tags: [check]
      summary: Enqueue an anonymous live check (async job + poll)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host]
              properties:
                host: { type: string, example: example.com }
      responses:
        "202":
          description: Job enqueued; poll the Location header. Recommended poll interval 2 s.
          headers:
            Location:
              schema: { type: string, example: /check/481529 }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckAccepted" }
        "200":
          description: Dedupe hit (domain- or job-side) — a cached done envelope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckEnvelope" }
        "400": { $ref: "#/components/responses/InvalidParameter" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /check/{id}:
    get:
      operationId: getCheck
      tags: [check]
      summary: Poll a live-check job to a terminal state
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: >
            Terminal (`done`/`failed`) responses cache `public, max-age=60`;
            in-flight polls are `no-store`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
...
    class:
      name: class
      in: query
      schema: { $ref: "#/components/schemas/Classification" }
    saint:
      name: saint
      in: query
      description: The only accepted value is true (saint ⊂ hero).
      schema: { type: boolean }
    country:
      name: country
      in: query
      x-go-name: CountryParam
      description: ISO code scope; composes with class + rank order.
      schema: { type: string }
    asn:
      name: asn
      in: query
      description: AS-number scope.
      schema: { type: integer, format: int64 }
    tld:
      name: tld
      in: query
      description: eTLD-suffix pivot; scope-required.
      schema: { type: string }
    provider:
      name: provider
      in: query
      x-go-name: ProviderParam
      description: dns_provider id pivot; scope-required as a query filter.
      schema: { type: integer, format: int64 }
    hosting:
      name: hosting
      in: query
      description: Hosting/CDN text-tag pivot; scope-required.
      schema: { type: string }
    flag:
      name: flag
      in: query
      description: class_flags residual; scope-required.
      schema: { $ref: "#/components/schemas/ClassFlag" }
...
    dimWWW:
      name: www
      in: query
      schema: { $ref: "#/components/schemas/IPv6Status" }
...